utils: Move buffer file writes into filesystem utilities
Durable filesystem replacement now has a low-level utility module. The
Buffer type still knows how to write paths, select working-tree modes, and
invoke the replacement sequence.

Buffers are imported throughout storage, command, and display code. Giving
that widely used data type responsibility for filesystem paths makes
low-level code depend on operating-system file-handling details that it
does not otherwise need.

This commit continues separating filesystem persistence responsibilities by
moving regular-file and working-tree output into a utility adapter that
accepts a core buffer as input.

The next commit will test the new dependency direction and prevent the
Buffer type from regaining filesystem responsibilities.
