+++ src/git_stage_batch/data/undo_restore.py
@@ -10,7 +10,7 @@ import tempfile
 from pathlib import Path
 from typing import Any
 
-from ..core.buffer import write_buffer_to_working_tree_path
+from ..utils.buffer_io import write_buffer_to_working_tree_path
 from ..utils.repository_buffers import load_git_blob_as_buffer
 from .undo_refs import list_restorable_refs
 from ..exceptions import CommandError
